added 404 error

+++ src/controllers/JobController.ts
@@ -187,6 +187,14 @@ export class JobController {
         });
         return;
       }
+      const job= await this.jobService.getJobById(jobRoleId);
+      if(job.length===0){
+        _res.status(404).json({
+          error: 'Not Found',
+          message: `Job with ID ${jobRoleId} not found`,
+        });
+        return;
+      }
       const response = await this.jobService.deleteJob(jobRoleId);
       if (response) {
         _res.status(200).json({
